+++ packages/react/src/components/Table/TableBody/TableBodyRow/TableDragHandle.jsx
@@ -1,6 +1,7 @@
 import { Draggable } from '@carbon/react/icons';
 import React, { forwardRef } from 'react';
 import PropTypes from 'prop-types';
+import { Tooltip } from '@carbon/react';
 
 import { settings } from '../../../../constants/Settings';
 
@@ -17,12 +18,25 @@ const propTypes = {
    * The ID of the table row this handle is it. This is the row that will be dragged by this handle.
    */
   rowId: PropTypes.string.isRequired,
+
+  /**
+   * If a drag operation is currently in progress. When true, the tooltip is hidden.
+   */
+  isDragging: PropTypes.bool.isRequired,
+
+  /**
+   * Tooltip text to display on hover.
+   */
+  tooltipText: PropTypes.string.isRequired,
 };
 
 /**
  * Drag handle image the user needs to click and drag to start a drag and drop operation.
  */
-const TableDragHandle = forwardRef(function TableDragHandle({ onStartDrag, rowId }, ref) {
+const TableDragHandle = forwardRef(function TableDragHandle(
+  { onStartDrag, rowId, isDragging, tooltipText },
+  ref
+) {
   return (
     <div
       className={`${iotPrefix}--table-drag-handle`}
@@ -34,7 +48,17 @@ const TableDragHandle = forwardRef(function TableDragHandle({ onStartDrag, rowId
         onStartDrag(e, rowId);
       }}
     >
-      <Draggable />
+      {isDragging ? (
+        // During drag, render icon without tooltip
+        <div className={`${iotPrefix}--table-drag-handle-icon`}>
+          <Draggable />
+        </div>
+      ) : (
+        // When not dragging, render with tooltip
+        <Tooltip label={tooltipText} autoAlign>
+          <Draggable />
+        </Tooltip>
+      )}
     </div>
   );
 });