+++ CHANGELOG.md
@@ -3,6 +3,53 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.6.4](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v5.6.3...v5.6.4) (2026-01-07)
+
+
+### Bug Fixes
+
+* add tooltip to drag handler #MAXUIF-3526 ([#4054](https://github.com/carbon-design-system/carbon-addons-iot-react/issues/4054)) ([69f6fed](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/69f6fed77aec8574ef902d39665038c19b5cffe4)), closes [#MAXUIF-3526](https://github.com/carbon-design-system/carbon-addons-iot-react/issues/MAXUIF-3526)
+
+
+
+
+
+## [5.6.3](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v5.6.2...v5.6.3) (2026-01-06)
+
+
+### Bug Fixes
+
+* add logging to release.sh script ([9ece291](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/9ece291470d22758ff907eaf31ea4304bf82554a))
+* enable OIDC in pipeline ([ff3f383](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/ff3f3838326025e930a7bbdf4c3469b9bd1884c1))
+* enable OIDC publishing ([4d71a68](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/4d71a68a92b296783ba4d6021d707a3d9fd926a7))
+* enable OIDC publishing ([7ca123a](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/7ca123abe5794222602483e955c33b46bb727d81))
+* testing OIDC publishing ([76a4600](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/76a4600f9637db27db1900329ab2cafeedf3b73f))
+* try to fix publishing ([16f1496](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/16f14969df2cf6ff2232403a42e9893a243ed23d))
+* updated to use OIDC ([ca09f17](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/ca09f17b727a0adab65ee8feef81550c37dfb4e7))
+
+
+
+
+
+## [5.6.2](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v5.6.1...v5.6.2) (2026-01-05)
+
+**Note:** Version bump only for package ibm-ai-applications
+
+
+
+
+
+## [5.6.1](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v5.6.0...v5.6.1) (2026-01-05)
+
+
+### Bug Fixes
+
+* revert yarn version changes ([#4055](https://github.com/carbon-design-system/carbon-addons-iot-react/issues/4055)) ([f3e4c6c](https://github.com/carbon-design-system/carbon-addons-iot-react/commit/f3e4c6ca234fcb3dd76615e9505c30ca0818a21f))
+
+
+
+
+
 # [5.6.0](https://github.com/carbon-design-system/carbon-addons-iot-react/compare/v5.5.2...v5.6.0) (2026-01-01)